// Web app/csi-mkd-premarital-app-FE/src/api/api-main-app/models/create-question-answers-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CreateQuestionAnswersDto {
  ageDifferenceImpact?: string | null;
  churchImportance?: string | null;
  definitionOfMarriage?: string | null;
  eldestYoungestScenario?: string | null;
  expectationsFromPartner?: string | null;
  familyBackground?: string | null;
  fearsAboutMarriage?: string | null;
  greatestAdjustment?: string | null;
  parentsHealthImpact?: string | null;
  premaritalRegistrationId: string;
  relationshipWithParentsInlaws?: string | null;
  timeWithPartner?: string | null;
  understandingAboutSex?: string | null;
  wishesConcerns?: string | null;
}